pyproject.toml
[project]
name = "zapis-stavy-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgi-correlation-id>=4.3.4",
    "fastapi>=0.128.0",
    "logtail-python>=0.3.4",
    "psycopg[binary,pool]>=3.3.2",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-json-logger>=4.0.0",
    "python-multipart>=0.0.21",
    "uvicorn[standard]>=0.38.0",
]
[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "ruff>=0.14.9",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "httpx>=0.28.1",
]


[tool.ruff]
line-length = 88
indent-width = 4

[tool.ruff.lint]
extend-select = [
    "E501", # line too long
    "N",    # PEP8 naming
    "R",    # RSE - flake8 raise, RET - flake8 return, RUF - ruff specific
    "I",    # isort
    "C901", # mccabe - complexity
    "PD",   # pandas
    "NPY",  # numpy
    "D"     # pydocstyle
]
ignore = [
    "D100", # missing docstring in public module
    "D104", # missing docstring in public package
    "D107", # missing docstring in __init__
]

[tool.ruff.lint.per-file-ignores]
"api/schemas/*" = ["D101"] # missing docstring in public class
"tests/*" = [
    "D102", # missing docstring in public method
    "D103", # missing docstring in public function
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
indent-style = "space"

[tool.mypy]
strict = true
explicit_package_bases = true
ignore_missing_imports = true
disable_error_code = ["no-any-return", "union-attr", "misc"]

[tool.pytest.ini_options]
markers = [
    "integration: marks online tests, that use connection to postgres test db.",
]
